requirements.txt
google-generativeai>=0.3.0
python-dotenv>=0.19.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
xmlrpc>=1.0.1
tabulate>=0.8.9
requests>=2.26.0
python-json-logger>=2.0.7
typing-extensions>=4.0.0 